Update README.md
Just in case someone is considering forking this fork, I was wanting to let them know this fork is terribly out of date and that I have been contributing to the main project for over a year now if you consider that all the work in this fork was pulled and squashed in to the main project.  No I did not get credit for this work since it was squashed.